requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
easyocr==1.7.1
Pillow==10.3.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
python-multipart==0.0.9
watchdog==4.0.1
pdf2image==1.17.0
PyMuPDF==1.24.3
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
reportlab==4.2.0
pytest==8.2.0
httpx==0.27.0
streamlit==1.35.0
loguru==0.7.2
colorama==0.4.6
requests==2.32.3
